+++ backend_spring/src/main/java/com/nigelkarunaratne/dbproject/repository/ManagerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ManagerRepository extends JpaRepository<Manager, Long> {
+public interface ManagerRepository extends JpaRepository<Manager, Integer> {
     // Possible TO do - add custom queries here? EX: List<Manager> findByExpertiseArea(String area);
 }
